[eclipse/xtext#1679] ported more code to java
