+++ org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/IXtextGeneratorLanguage.java
@@ -1,11 +1,11 @@
-/**
+/*******************************************************************************
  * Copyright (c) 2015, 2017 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
- */
+ *******************************************************************************/
 package org.eclipse.xtext.xtext.generator;
 
 import java.util.List;
@@ -19,29 +19,30 @@ import org.eclipse.xtext.xtext.generator.model.StandaloneSetupAccess;
  * 
  * @noimplement This interface should not be implemented by clients.
  */
-@SuppressWarnings("all")
 public interface IXtextGeneratorLanguage {
-  Grammar getGrammar();
-  
-  List<String> getFileExtensions();
-  
-  StandaloneSetupAccess getRuntimeGenSetup();
-  
-  GuiceModuleAccess getRuntimeGenModule();
-  
-  GuiceModuleAccess getIdeGenModule();
-  
-  GuiceModuleAccess getEclipsePluginGenModule();
-  
-  GuiceModuleAccess getIdeaGenModule();
-  
-  GuiceModuleAccess getWebGenModule();
-  
-  ResourceSet getResourceSet();
-  
-  /**
-   * @deprecated Use {@link CodeConfig#isPreferXtendStubs()} instead
-   */
-  @Deprecated
-  boolean isGenerateXtendStubs();
+	
+	Grammar getGrammar();
+	
+	List<String> getFileExtensions();
+	
+	StandaloneSetupAccess getRuntimeGenSetup();
+	
+	GuiceModuleAccess getRuntimeGenModule();
+	
+	GuiceModuleAccess getIdeGenModule();
+	
+	GuiceModuleAccess getEclipsePluginGenModule();
+	
+	GuiceModuleAccess getIdeaGenModule();
+	
+	GuiceModuleAccess getWebGenModule();
+	
+	ResourceSet getResourceSet();
+	
+	/**
+	 * @deprecated Use {@link CodeConfig#isPreferXtendStubs()} instead
+	 */
+	@Deprecated
+	boolean isGenerateXtendStubs();
+	
 }
