+++ org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/model/annotations/IClassAnnotation.java
@@ -1,11 +1,11 @@
-/**
- * Copyright (c) 2015, 2017 itemis AG (http://www.itemis.eu) and others.
+/*******************************************************************************
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
- */
+ *******************************************************************************/
 package org.eclipse.xtext.xtext.generator.model.annotations;
 
 import java.lang.annotation.Repeatable;
@@ -13,7 +13,7 @@ import org.eclipse.xtext.xtext.generator.XtextGenerator;
 import org.eclipse.xtext.xtext.generator.model.JavaFileAccess;
 import org.eclipse.xtext.xtext.generator.model.TypeReference;
 
-/**
+/** 
  * Class annotations can be added to the {@link XtextGenerator} workflow component in order
  * to configure specific Java annotations to be added to each generated class.
  * <br/><br/>
@@ -23,21 +23,22 @@ import org.eclipse.xtext.xtext.generator.model.TypeReference;
  * 	<li>If the underlying Annotation is NOT {@link Repeatable} then {@link Object#equals(Object)} should return true for all instances and {@link Object#hashCode()} should return the same value for all instances</li>
  * </ul>
  */
-@SuppressWarnings("all")
 public interface IClassAnnotation {
-  /**
-   * Convert the class annotation to a string suitable for use in Java code generation.
-   */
-  CharSequence generate();
-  
-  /**
-   * Determine whether this annotation should be applied to the given Java file.
-   */
-  boolean appliesTo(final JavaFileAccess javaFile);
-  
-  /**
-   * Return the qualified name of the annotation interface for use in import declarations,
-   * or {@code null} if no import is required.
-   */
-  TypeReference getAnnotationImport();
+	
+	/**
+	 * Convert the class annotation to a string suitable for use in Java code generation.
+	 */
+	CharSequence generate();
+	
+	/**
+	 * Determine whether this annotation should be applied to the given Java file.
+	 */
+	boolean appliesTo(JavaFileAccess javaFile);
+
+	/** 
+	 * Return the qualified name of the annotation interface for use in import declarations,
+	 * or {@code null} if no import is required.
+	 */
+	TypeReference getAnnotationImport();
+
 }
