+++ org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/model/project/IBundleProjectConfig.java
@@ -1,22 +1,25 @@
 /*******************************************************************************
- * Copyright (c) 2015, 2017 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.xtext.xtext.generator.model.project
+package org.eclipse.xtext.xtext.generator.model.project;
 
-import org.eclipse.xtext.xtext.generator.model.IXtextGeneratorFileSystemAccess
+import org.eclipse.xtext.xtext.generator.model.ManifestAccess;
+import org.eclipse.xtext.xtext.generator.model.PluginXmlAccess;
 
 /**
- * Configuration of the web project.
+ * Configuration of subprojects that can be used as Eclipse bundles.
  * 
  * @noimplement This interface should not be implemented by clients.
  */
-interface IWebProjectConfig extends ISubProjectConfig {
+public interface IBundleProjectConfig extends ISubProjectConfig {
 	
-	def IXtextGeneratorFileSystemAccess getAssets()
+	ManifestAccess getManifest();
+	
+	PluginXmlAccess getPluginXml();
 	
 }
