+++ org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/model/project/IWebProjectConfig.java
@@ -1,25 +1,22 @@
 /*******************************************************************************
- * Copyright (c) 2015, 2017 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.xtext.xtext.generator.model.project
+package org.eclipse.xtext.xtext.generator.model.project;
 
-import org.eclipse.xtext.xtext.generator.model.ManifestAccess
-import org.eclipse.xtext.xtext.generator.model.PluginXmlAccess
+import org.eclipse.xtext.xtext.generator.model.IXtextGeneratorFileSystemAccess;
 
 /**
- * Configuration of subprojects that can be used as Eclipse bundles.
+ * Configuration of the web project.
  * 
  * @noimplement This interface should not be implemented by clients.
  */
-interface IBundleProjectConfig extends ISubProjectConfig {
+public interface IWebProjectConfig extends ISubProjectConfig {
 	
-	def ManifestAccess getManifest()
-	
-	def PluginXmlAccess getPluginXml()
+	IXtextGeneratorFileSystemAccess getAssets();
 	
 }
