+++ org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/serializer/InjectableRuleNames.java
@@ -1,11 +1,11 @@
-/**
- * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
+/*******************************************************************************
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
- */
+ *******************************************************************************/
 package org.eclipse.xtext.xtext.generator.serializer;
 
 import com.google.inject.Inject;
@@ -13,14 +13,15 @@ import com.google.inject.Singleton;
 import org.eclipse.xtext.Grammar;
 import org.eclipse.xtext.xtext.RuleNames;
 
-/**
+/** 
  * @author Sebastian Zarnekow - Initial contribution and API
  */
-@Singleton
-@SuppressWarnings("all")
+@Singleton 
 public class InjectableRuleNames extends RuleNames {
-  @Inject
-  public InjectableRuleNames(final Grammar grammar) {
-    super(grammar, false);
-  }
+	
+	@Inject
+	public InjectableRuleNames(Grammar grammar) {
+		super(grammar, false);
+	}
+	
 }
