+++ org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/util/SyntheticTerminalDetector.java
@@ -1,16 +1,18 @@
 /** 
- * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.xtext.xtext.generator.util
+package org.eclipse.xtext.xtext.generator.util;
 
-import org.eclipse.xtext.Keyword
-import org.eclipse.xtext.TerminalRule
-import org.eclipse.xtext.xtext.generator.parser.antlr.AntlrGrammarGenUtil
+import org.eclipse.xtext.Keyword;
+import org.eclipse.xtext.TerminalRule;
+import org.eclipse.xtext.xtext.generator.parser.antlr.AntlrGrammarGenUtil;
+
+import com.google.common.base.Objects;
 
 /** 
  * Helper to identify synthetic terminal rules.
@@ -19,19 +21,20 @@ import org.eclipse.xtext.xtext.generator.parser.antlr.AntlrGrammarGenUtil
  * 
  * @author Sebastian Zarnekow - Initial contribution and API
  */
-class SyntheticTerminalDetector {
+public class SyntheticTerminalDetector {
 	
 	/** 
 	 * Answers {@code true} if the given terminal rule is synthetic. That is,
 	 * the tokens for this rule will not be produced by the generated Antlr lexer
 	 * but manually in a custom token source.
 	 */
-	def boolean isSyntheticTerminalRule(TerminalRule rule) {
-		if (rule.alternatives instanceof Keyword) {
-			var String value = (rule.alternatives as Keyword).value
-			return 'synthetic:' + AntlrGrammarGenUtil.getOriginalElement(rule).name == value
+	public boolean isSyntheticTerminalRule(TerminalRule rule) {
+		if (rule.getAlternatives() instanceof Keyword) {
+			String value = ((Keyword) rule.getAlternatives()).getValue();
+			String sytheticValue = "synthetic:" + AntlrGrammarGenUtil.getOriginalElement(rule).getName();
+			return Objects.equal(sytheticValue, value);
 		}
-		return false
+		return false;
 	}
 	
 }
