--- org.eclipse.xtext.xtext.generator/xtend-gen/org/eclipse/xtext/xtext/generator/model/project/IBundleProjectConfig.java
@@ -1,25 +1,0 @@
-/**
- * Copyright (c) 2015, 2017 itemis AG (http://www.itemis.eu) and others.
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- * 
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.xtext.xtext.generator.model.project;
-
-import org.eclipse.xtext.xtext.generator.model.ManifestAccess;
-import org.eclipse.xtext.xtext.generator.model.PluginXmlAccess;
-import org.eclipse.xtext.xtext.generator.model.project.ISubProjectConfig;
-
-/**
- * Configuration of subprojects that can be used as Eclipse bundles.
- * 
- * @noimplement This interface should not be implemented by clients.
- */
-@SuppressWarnings("all")
-public interface IBundleProjectConfig extends ISubProjectConfig {
-  ManifestAccess getManifest();
-  
-  PluginXmlAccess getPluginXml();
-}
