--- org.eclipse.xtext.xtext.generator/xtend-gen/org/eclipse/xtext/xtext/generator/model/project/IWebProjectConfig.java
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) 2015, 2017 itemis AG (http://www.itemis.eu) and others.
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- * 
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.xtext.xtext.generator.model.project;
-
-import org.eclipse.xtext.xtext.generator.model.IXtextGeneratorFileSystemAccess;
-import org.eclipse.xtext.xtext.generator.model.project.ISubProjectConfig;
-
-/**
- * Configuration of the web project.
- * 
- * @noimplement This interface should not be implemented by clients.
- */
-@SuppressWarnings("all")
-public interface IWebProjectConfig extends ISubProjectConfig {
-  IXtextGeneratorFileSystemAccess getAssets();
-}
